feat: add new processing paradigm
This processing strategy is designed to allow for any and all existing/future files to be processed by markdown magic, while leaving the door open for future processing in the future.

In short, we deprecated the custom `[name]` style template variables in favor of the open source "handlebars" syntax `{{ name }}` as it supports custom transforms out of the box (on top of being widely used).